requirements.txt
chainlit
openai
langchain
python-dotenv
psycopg2
python-dotenv
chromadb
googlemaps